An attempt to build, extend and keep graph.
This is an attempt to keep the graph from emptying and reloading
constantly. Since we refocus the graph with contextual click events,
and throw out old data when doing so, the graph jumps often. This
tries to minimize this behaviour by keeping and merging old and new
loaded graph data.