Use nav links to move between selected items.
Use nav links to move between selected items instead of keeping track
of click events via state. This way, we can decouple the origin and
destination of navitaional events. I.e. selecting an item in a view
can be done outside of that view simply by clicking a navigational elm
like a link.