Moving selected item state out of graph view.
In order to use nav links to move between selected items instead of
keeping track of click events via state, we have to move the state of
the selected center out of the graph view, and instead keep track of
it in parent views and pass it in to the graph. The parent view should
simply pass in the intended selected center and let the graph move.