Attempt to use as much space as possible for Graph View.
Attempt to use as much space as possible for Graph View by setting the
size of the Graph view to use all available space within the node to
render in. This seems to be best done by setting the width and height
to match a known ref node within the view.